// src/app/breakpoint-detector/breakpoint-detector.interface.ts
import { Breakpoint } from './breakpoint-detector-breakpoint.interface';

export interface BreakpointDetectorConfig {
  BREAKPOINT_MOBILE: Breakpoint;
  BREAKPOINT_TABLET_PORTRAIT: Breakpoint;
  BREAKPOINT_TABLET_LANDSCAPE: Breakpoint;
  BREAKPOINT_MEDIUM_DESKTOP: Breakpoint;
  BREAKPOINT_LARGE_DESKTOP: Breakpoint;
  BREAKPOINT_EXTRA_LARGE_DESKTOP: Breakpoint;
}